+++ folio-spring-system-user/src/main/java/org/folio/spring/exception/SystemUserAuthorizationException.java
@@ -1,5 +1,6 @@
 package org.folio.spring.exception;
 
+@Deprecated(since = "10.0.0", forRemoval = true)
 public class SystemUserAuthorizationException extends RuntimeException {
   public SystemUserAuthorizationException(String message) {
     super(message);