+++ folio-spring-system-user/src/main/java/org/folio/spring/model/SystemUser.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.With;
 
 @Builder
+@Deprecated(since = "10.0.0", forRemoval = true)
 public record SystemUser(String username, String okapiUrl, String tenantId,
                          @With UserToken token, @With String userId) {
 }